services/work-package-service/requirements-dev.txt
# requirements for development and testing this service

-r requirements-dev-common.txt

# additional requirements can be listed here
testcontainers[kafka,mongo]==3.4.1
pytest_httpx==0.21.3
